Create audit files
This code runs through the current cleaning code and creates a separate stata file. This will make it easier to compare the spreadsheets and the data in stata. I already found some issues when looping and I used "capture" to bypass errors. Parker's files created a lot of issues. 
I will have to remove the "capture" command to see what files are causing issues.